requirements.txt
# requirements.py
pip==25.0.1
aiogram==3.18.0
httpx==0.28.1
asyncpg==0.30.0
python-dotenv==1.0.1
openai==1.66.3
gspread==6.2.0
oauth2client==4.1.3
requests==2.32.3
pytest==8.3.5
psycopg2-binary==2.9.10
SQLAlchemy==2.0.39
alembic==1.15.1



google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0

protobuf==5.29.4
apscheduler==3.11.0
pytest-asyncio==0.25.3
pytz==2025.2
yandex-cloud-ml-sdk==0.7.0










# Mako-1.3.9 MarkupSafe-3.0.2 SQLAlchemy-2.0.38 aiofiles-24.1.0 aiogram-3.18.0 aiohappyeyeballs-2.5.0 aiohttp-3.11.13 aiosignal-1.3.2 alembic-1.15.1 annotated-types-0.7.0 anyio-4.8.0 asyncpg-0.30.0 attrs-25.1.0 cachetools-5.5.2 certifi-2025.1.31 charset-normalizer-3.4.1 distro-1.9.0 frozenlist-1.5.0 google-auth-2.38.0 google-auth-oauthlib-1.2.1 gspread-6.2.0 h11-0.14.0 httpcore-1.0.7 httplib2-0.22.0 httpx-0.28.1 idna-3.10 iniconfig-2.0.0 jiter-0.9.0 magic-filter-1.0.12 multidict-6.1.0 oauth2client-4.1.3 oauthlib-3.2.2 openai-1.65.5 packaging-24.2 pluggy-1.5.0 propcache-0.3.0 psycopg2-binary-2.9.10 pyasn1-0.6.1 pyasn1-modules-0.4.1 pydantic-2.10.6 pydantic-core-2.27.2 pyparsing-3.2.1 pytest-8.3.5 python-dotenv-1.0.1 requests-2.32.3 requests-oauthlib-2.0.0 rsa-4.9 six-1.17.0 sniffio-1.3.1 tqdm-4.67.1 typing-extensions-4.12.2 urllib3-2.3.0 yarl-1.18.3

